fix id not found

+++ app/Services/Admin/BranchService.php
@@ -3,6 +3,7 @@
 namespace App\Services\Admin;
 
 use App\Models\Branch;
+use Illuminate\Validation\ValidationException;
 
 class BranchService
 {
@@ -14,26 +15,36 @@ class BranchService
         //
     }
 
-
     public function getBranches()
     {
         return Branch::all();
     }
-
     public function createBranch(array $data)
     {
         return Branch::create($data);
     }
-
     public function viewBranch($id)
     {
-        return Branch::findOrFail($id);
-    }
+        $branch = Branch::find($id);
 
+        if (!$branch) {
+            throw ValidationException::withMessages([
+                'payment' => 'Branch id not found.',
+            ]);
+        }
+
+        return $branch;
+    }
     public function editBranch($id, array $data)
     {
-        $branch = Branch::findOrFail($id);
-        
+        $branch = Branch::find($id);
+
+        if (!$branch) {
+            throw ValidationException::withMessages([
+                'message' => 'Branch id not found.',
+            ]);
+        }
+
         $branch->name = $data['name'] ?? $branch->name;
         $branch->location = $data['location'] ?? $branch->location;
         $branch->latitude = $data['latitude'] ?? $branch->latitude;
@@ -41,14 +52,21 @@ class BranchService
         $branch->country = $data['country'] ?? $branch->country;
         $branch->working_hour = $data['working_hour'] ?? $branch->working_hour;
         $branch->save();
-        
+
         return $branch;
     }
-
     public function deleteBranch($id)
     {
-        $branch = Branch::findOrFail($id);
+        $branch = Branch::find($id);
+
+        if (!$branch) {
+            throw ValidationException::withMessages([
+                'message' => 'Branch id not found.',
+            ]);
+        }
+
         $branch->delete();
+
         return true;
     }
 }
